# internal/model/model.go
package model

type WebResponse[T any] struct {
	Data T `json:"data"`
}

type PageResponse[T any] struct {
	Data   []T           `json:"data"`
	Paging *PageMetadata `json:"paging"`
}

type PageMetadata struct {
	Page      int   `json:"page"`
	Size      int   `json:"size"`
	TotalItem int64 `json:"total_item"`
	TotalPage int64 `json:"total_page"`
}

type ErrorResponse struct {
	Errors string `json:"errors"`
}